+++ b4n/kube/resources/mod.rs
@@ -1,6 +1,6 @@
 pub use self::data::*;
 pub use self::observer::ResourceObserver;
-pub use self::resource::{ResourceFilterContext, ResourceItem};
+pub use self::resource::{ColumnsLayout, ResourceFilterContext, ResourceItem};
 pub use self::resource_data::{ResourceData, ResourceValue};
 pub use self::resources_list::ResourcesList;